+++ src/SampleApps/SampleApp.Angular/Program.cs
@@ -33,7 +33,7 @@ if (!app.Environment.IsDevelopment())
 	app.UseSpaStaticFiles();
 }
 
-app.UseSimplifyWebNonTerminal();
+app.UseSimplifyWebNonTerminal(true);
 
 BootstrapperFactory.ContainerProvider.Verify();