+++ src/SampleApps/SampleApp.WindowsServiceHosted/Startup.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Simplify.Web;
-using Simplify.Web.Bootstrapper;
 
 namespace SampleApp.WindowsServiceHosted;
 
@@ -12,8 +11,6 @@ public class Startup
 		if (env.IsDevelopment())
 			app.UseDeveloperExceptionPage();
 
-		app.UseSimplifyWebWithoutRegistrations();
-
-		BootstrapperFactory.ContainerProvider.Verify();
+		app.UseSimplifyWeb();
 	}
 }