+++ src/Simplify.Web/Bootstrapper/Setup/BaseBootstrapperConfiguration.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Linq;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Simplify.DI;
-using Simplify.System;
 
 namespace Simplify.Web.Bootstrapper.Setup;
 
@@ -14,17 +11,6 @@ public partial class BaseBootstrapper
 	/// <summary>
 	/// Registers the configuration.
 	/// </summary>
-	public virtual void RegisterConfiguration()
-	{
-		if (TypesToExclude.Contains(typeof(IConfiguration)))
-			return;
-
-		var environmentName = Environment.GetEnvironmentVariable(ApplicationEnvironment.EnvironmentVariableName);
-
-		var builder = new ConfigurationBuilder()
-			.AddJsonFile("appsettings.json", true)
-			.AddJsonFile($"appsettings.{environmentName}.json", true);
-
-		BootstrapperFactory.ContainerProvider.Register<IConfiguration>(r => builder.Build(), LifetimeType.Singleton);
-	}
+	private static void RegisterConfiguration(IConfiguration configuration) =>
+		BootstrapperFactory.ContainerProvider.Register(r => configuration, LifetimeType.Singleton);
 }