+++ src/Simplify.Web/Bootstrapper/Setup/BaseBootstrapperSettings.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using Simplify.DI;
+﻿using Simplify.DI;
 using Simplify.Web.Settings;
 
 namespace Simplify.Web.Bootstrapper.Setup;
@@ -12,11 +11,6 @@ public partial class BaseBootstrapper
 	/// <summary>
 	/// Registers the Simplify.Web settings.
 	/// </summary>
-	public virtual void RegisterSimplifyWebSettings()
-	{
-		if (TypesToExclude.Contains(typeof(ISimplifyWebSettings)))
-			return;
-
-		BootstrapperFactory.ContainerProvider.Register<ISimplifyWebSettings, SimplifyWebSettings>(LifetimeType.Singleton);
-	}
+	private static void RegisterSimplifyWebSettings(ISimplifyWebSettings settings) =>
+		BootstrapperFactory.ContainerProvider.Register(r => settings, LifetimeType.Singleton);
 }