+++ src/Simplify.Web/Bootstrapper/Setup/BaseBootstrapperViews.cs
@@ -17,6 +17,6 @@ public partial class BaseBootstrapper
 		if (TypesToExclude.Contains(typeof(IViewFactory)))
 			return;
 
-		BootstrapperFactory.ContainerProvider.Register<IViewFactory, ViewFactory>();
+		BootstrapperFactory.ContainerProvider.Register<IViewFactory>(r => new ViewFactory(r));
 	}
 }